Try fix compare in docs
