+++ readme/Footer.scalatex
@@ -96,9 +96,9 @@
         @lnk("an sbt plug-in", "https://github.com/ThoughtWorksInc/sbt-ammonite-classpath")
         to export the classpath of an sbt project to Ammonite Script.
       @li
-        @link("Elastic", "https://www.elastic.co/")
+        @lnk("Elastic", "https://www.elastic.co/")
         uses Ammonite as a management repl inside their
-        @link("Elastic Cloud Enterprise", "https://www.elastic.co/products/ece")
+        @lnk("Elastic Cloud Enterprise", "https://www.elastic.co/products/ece")
         product.
 
   @sect{Scaladoc}
@@ -107,15 +107,15 @@
 
     @ul
       @li
-        @a("Ammonite-Ops", href:="https://lihaoyi.github.io/Ammonite/api/ops/index.html")
+        @a("Ammonite-Ops", href:="https://ammonite.io/api/ops/index.html")
       @li
-        @a("Ammonite-Terminal", href:="https://lihaoyi.github.io/Ammonite/api/terminal/index.html")
+        @a("Ammonite-Terminal", href:="https://ammonite.io/api/terminal/index.html")
       @li
-        @a("Ammonite-Repl", href:="https://lihaoyi.github.io/Ammonite/api/repl/index.html")
+        @a("Ammonite-Repl", href:="https://ammonite.io/api/repl/index.html")
       @li
-        @a("Ammonite-Shell", href:="https://lihaoyi.github.io/Ammonite/api/shell/index.html")
+        @a("Ammonite-Shell", href:="https://ammonite.io/api/shell/index.html")
       @li
-        @a("Ammonite-Sshd", href:="https://lihaoyi.github.io/Ammonite/api/sshd/index.html")
+        @a("Ammonite-Sshd", href:="https://ammonite.io/api/sshd/index.html")
 
     @p
       Although it's best to read the documentation on this page to learn how
